split up TodoList to smaller components
