+++ README.md
@@ -6,9 +6,12 @@ The goal of this project was to build a todo app using Redux store.
 
 - Responsive UI for mobile, tablet and desktop
 - Redux hooks: useSelector, useDispatch
+- Ternary operators to handle empty states
+- Timestap added using moment.js
+- Summary componant that counts completed todos and uncompleted
 
 ## Production process
 
-- The UI was sketched on Figma.
+- The UI was sketched in Figma.
 
 ## View it live
