+++ code/public/index.html
@@ -4,15 +4,17 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
+    <meta property="og:title" content="Daily help - todo-app" />
 
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
+    <meta property="og:description" content="A todo-app for your daily tasks. Keep it clean and feel free!" />
+
+    <meta property="og:type" content="website" />
+
+    <meta property="og:url" content="#" />
+
+    <meta property="og:image" content="https://aspen-dailystandup.netlify.app/images/ 
+
+..." />
     <title>DAILY HELP(!)</title>
   </head>
