+++ code/src/App.js
@@ -3,7 +3,6 @@ import { Provider } from 'react-redux';
 import { combineReducers, createStore } from '@reduxjs/toolkit';
 
 import todos from './reducers/todos';
-import Header from './components/Header';
 import AddTodo from './components/AddTodos';
 import TodoList from './components/TodoList';
 import Summary from './components/Summary';
