+++ code/src/TODO
@@ -15,7 +15,7 @@ Add accordion to show filtered todos?
 https://react-bootstrap.github.io/components/accordion/
 
 Styled components?
-Lägg till bilden i assets
+Lägg till bilden i assets✅
 bryt rad - se happy thoughts✅
 skicka på enter
 CSS i todolist✅
@@ -23,5 +23,10 @@ Checkbox synlig✅
 Delete all knapp✅
 Date i grått eller opacity
 Add knapp större ✅
-centrera summary
-Ta bort header?
+centrera summary ✅
+Ta bort header? ✅
+Publicera netlify
+Addera kommentarer
+Ta bort console log
+KOmplettera meta tags
+Komplettera readme
