--- code/src/components/Header.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import './header.css';
-
-export const Header = () => {
-  return (
-    <header>
-      <h1>Daily help(!)</h1>
-    </header>
-  );
-};
-
-export default Header;
