+++ code/src/components/TodoList.js
@@ -1,24 +1,14 @@
 import React from 'react';
-import { FaTrash } from 'react-icons/fa';
-import './todolist.css';
-import moment from 'moment';
-
 import { useDispatch, useSelector } from 'react-redux';
+
 import todos from '../reducers/todos';
+import CompletedTodos from './CompletedTodos';
+import UnCompletedTodos from './UnCompletedTodos';
+import './todolist.css';
 
 const TodoList = () => {
   const items = useSelector((store) => store.todos.items);
-
   const dispatch = useDispatch();
-
-  const onToggleTodo = (id) => {
-    dispatch(todos.actions.toggleTodo(id));
-  };
-
-  // delete one todo
-  const onDeleteTodo = (id) => {
-    dispatch(todos.actions.deleteTodo(id));
-  };
 
   // complete all todo
   const onClickCompleteAll = () => {
@@ -32,30 +22,10 @@ const TodoList = () => {
   // filter todo
   const unCompletedTodos = items.filter((item) => !item.isComplete);
   const completeTodos = items.filter((item) => item.isComplete);
-  console.log('UNCOMPL', unCompletedTodos);
-  const date = new Date();
 
   return (
     <>
-      {unCompletedTodos.map((item) => (
-        <div className="todo-card" key={item.id}>
-          <label htmlFor="checkbox" />
-          <input
-            className="round"
-            type="checkbox"
-            checked={item.isComplete}
-            onChange={() => onToggleTodo(item.id)}
-            id="checkbox"
-          />
-
-          <p className="todo">{item.text}</p>
-          <span className="date">{moment(date).format('ll')}</span>
-          <FaTrash
-            className="delete-button"
-            onClick={() => onDeleteTodo(item.id)}
-          />
-        </div>
-      ))}
+      <UnCompletedTodos />
       {unCompletedTodos.length > 0 && (
         <div className="button-wrapper">
           <button
@@ -67,26 +37,7 @@ const TodoList = () => {
           </button>
         </div>
       )}
-      {completeTodos.map((item) => (
-        <div className="todo-card completed" key={item.id}>
-          <label htmlFor="checkbox" />
-          <input
-            className="round"
-            type="checkbox"
-            checked={item.isComplete}
-            onChange={() => onToggleTodo(item.id)}
-            id="checkbox"
-          />
-
-          <p className="todo">{item.text}</p>
-
-          <span className="date">{moment(date).format('ll')}</span>
-          <FaTrash
-            className="delete-button"
-            onClick={() => onDeleteTodo(item.id)}
-          />
-        </div>
-      ))}
+      <CompletedTodos />
       {completeTodos.length > 0 && (
         <div className="button-wrapper">
           <button
