--- code/src/components/header.css
@@ -1,29 +1,0 @@
-header {
-  position: absolute;
-  top: 0%;
-  font-family: 'Autour One', cursive;
-
-  padding: 0;
-  margin: 0;
-}
-
-h1 {
-  margin: 5px 0 0 5px;
-  text-transform: uppercase;
-  color: var(--teal);
-  opacity: 50%;
-  font-size: 10px;
-}
-/* @media screen and (min-width: 768px) {
-  header {
-    font-size: 28px;
-    padding-top: 50px;
-  }
-}
-@media screen and (min-width: 992px) {
-  header {
-    font-size: 40px;
-    border-bottom: 1px solid black;
-    padding: 0 0 5% 0;
-  }
-}  */
