+++ code/src/components/todolist.css
@@ -50,7 +50,7 @@
   border-radius: 50%;
   cursor: pointer;
   -webkit-appearance: none;
-  outline: none;
+  border: 1px solid var(--grey);
   cursor: pointer;
 }
